new: support building ASYNC cap on osx and windows.
