+++ go-worker/go.mod
@@ -10,12 +10,12 @@ require (
 	github.com/containers/podman/v5 v5.4.0
 	github.com/docker/docker v27.5.1+incompatible
 	github.com/falcosecurity/plugin-sdk-go v0.7.4
+	github.com/fsnotify/fsnotify v1.8.0
 	github.com/google/uuid v1.6.0
 	github.com/opencontainers/runtime-spec v1.2.0
 	github.com/stretchr/testify v1.10.0
 	k8s.io/cri-api v0.32.0-alpha.0
 	k8s.io/cri-client v0.31.3
-	k8s.io/utils v0.0.0-20240711033017-18e509b52bc8
 )
 
 require (
@@ -61,7 +61,6 @@ require (
 	github.com/docker/go-connections v0.5.0 // indirect
 	github.com/docker/go-units v0.5.0 // indirect
 	github.com/felixge/httpsnoop v1.0.4 // indirect
-	github.com/fsnotify/fsnotify v1.8.0 // indirect
 	github.com/go-jose/go-jose/v4 v4.0.4 // indirect
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
@@ -180,6 +179,7 @@ require (
 	k8s.io/client-go v0.31.3 // indirect
 	k8s.io/component-base v0.31.3 // indirect
 	k8s.io/klog/v2 v2.130.1 // indirect
+	k8s.io/utils v0.0.0-20240711033017-18e509b52bc8 // indirect
 	sigs.k8s.io/yaml v1.4.0 // indirect
 	tags.cncf.io/container-device-interface v0.8.0 // indirect
 	tags.cncf.io/container-device-interface/specs-go v0.8.0 // indirect
