+++ go-worker/pkg/container/engine.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	"github.com/FedeDP/container-worker/pkg/config"
 	"github.com/FedeDP/container-worker/pkg/event"
-	"k8s.io/utils/inotify"
+	"github.com/fsnotify/fsnotify"
 	"net/url"
 	"os"
 	"path/filepath"
@@ -47,59 +47,6 @@ func (t engineType) ToCTValue() int {
 type engineGenerator func(context.Context, string) (Engine, error)
 type EngineGenerator func(ctx context.Context) (Engine, error)
 
-type EngineInotifier struct {
-	watcher           *inotify.Watcher
-	watcherGenerators map[string]EngineGenerator
-}
-
-func (e *EngineInotifier) Listen() <-chan *inotify.Event {
-	if e.watcher == nil {
-		return nil
-	}
-	return e.watcher.Event
-}
-
-func (e *EngineInotifier) Process(ctx context.Context, val interface{}) Engine {
-	ev, _ := val.(*inotify.Event)
-	if cb, ok := e.watcherGenerators[ev.Name]; ok {
-		_ = e.watcher.RemoveWatch(filepath.Dir(ev.Name))
-		engine, _ := cb(ctx)
-		return engine
-	} else {
-		// If the new created path is a folder, check if
-		// it is a subpath of any watcherGenerator socket,
-		// and eventually add a new watch, removing the old one
-		fileInfo, err := os.Stat(ev.Name)
-		if err != nil || !fileInfo.IsDir() {
-			return nil
-		}
-		for socket, cb := range e.watcherGenerators {
-			if strings.HasPrefix(socket, ev.Name) {
-				// Remove old watch
-				_ = e.watcher.RemoveWatch(filepath.Dir(ev.Name))
-				// It may happen that the actual socket has already been created.
-				// Check it and if it is not created yet, add a new inotify watcher.
-				if _, statErr := os.Stat(socket); os.IsNotExist(statErr) {
-					// Add new watch
-					_ = e.watcher.AddWatch(ev.Name, inotify.InCreate|inotify.InIsdir)
-					break
-				} else {
-					engine, _ := cb(ctx)
-					return engine
-				}
-			}
-		}
-	}
-	return nil
-}
-
-func (e *EngineInotifier) Close() {
-	if e.watcher == nil {
-		return
-	}
-	_ = e.watcher.Close()
-}
-
 // Hooked up by each engine through init()
 var engineGenerators = make(map[engineType]engineGenerator)
 
@@ -123,16 +70,16 @@ func Generators() ([]EngineGenerator, *EngineInotifier, error) {
 			if _, statErr := os.Stat(socket); os.IsNotExist(statErr) {
 				// Does not exist; emplace back an inotify listener
 				if engineNotifier.watcher == nil {
-					engineNotifier.watcher, _ = inotify.NewWatcher()
+					engineNotifier.watcher, _ = fsnotify.NewWatcher()
 				}
 				if engineNotifier.watcher != nil {
 					dir := filepath.Dir(socket)
-					err := engineNotifier.watcher.AddWatch(dir, inotify.InCreate)
+					err := engineNotifier.watcher.Add(dir)
 					if err != nil {
 						// Try to attach watcher to parent dir
 						// eg: /run/user for podman, /run/ for crio, and so on
 						dir = filepath.Dir(dir)
-						err = engineNotifier.watcher.AddWatch(dir, inotify.InCreate)
+						err = engineNotifier.watcher.Add(dir)
 					}
 					if err == nil {
 						engineNotifier.watcherGenerators[socket] = func(ctx context.Context) (Engine, error) {
