+++ go-worker/pkg/container/podman.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package container
 
 import (
