+++ go-worker/pkg/container/podman_others.go
@@ -1,0 +1,5 @@
+//go:build !linux
+
+package container
+
+const typePodman engineType = "podman"
