+++ go-worker/pkg/container/podman_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package container
 
 import (
